requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mangum==0.17.0
mysql-connector-python==8.2.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6